RapidCompetitions/requirements.txt
APScheduler==3.10.4
asgiref==3.6.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==4.2
django-apscheduler==0.6.2
django-cors-headers==4.3.1
django-environ==0.11.2
django-jazzmin==3.0.0
djangorestframework==3.15.1
gunicorn==20.1.0
idna==3.7
oauthlib==3.2.2
pillow==10.3.0
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2024.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.4.3
stripe==9.11.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
urllib3==2.2.1
whitenoise==6.4.0
